+++ src/main/java/company/entity/ObjectiveTest.java
@@ -8,10 +8,10 @@ import java.util.List;
 
 @Data
 public class ObjectiveTest {
-    int id;
+    long id;
     String testName;
 //    ObjectId subject_id;
-    int subject_id;
+    long subject_id;
     String poster;
     List<Question> questions = new ArrayList<>();
 }
